refactor: batch-1 modularity and compliance fixes from six-lens audit
Rendering: deduplicate _truncate_error_context/_ERROR_CONTEXT_LIMIT into
_output_text.py (was byte-identical in three renderers); de-shadow
'import re' by four 'except RenderingError as re:' bindings in
_combined_exports.py; interpolate _TIMEOUT_SECONDS into the DOCX/EPUB
timeout messages; fix ebook_stage docstring stage number (07 -> 11,
matching stage_11_ebook.py); promote the MathJax integrity/font/prefix
constants to public names in _web_assets with compat aliases.

Core: split execution_boundary.py - the ~330-line process-tree
termination region (POSIX descendant scans, macOS KERN_PROCARGS2 ctypes
parsing, bounded-run cleanup) moves to _bounded_kill.py with re-exports,
isolating the security-critical platform-kill logic for review.
Annotate core/cli.py as a compatibility/test surface; add
name-collision decision rules to both progress module docstrings.

Tests: relocate the 165-line Ollama session fixture trio + autouse hook
from the root conftest to tests/infra_tests/llm/conftest.py (root
conftest 347 -> 181 lines).

Exemplars: add source=["src"] to template_autoresearch_project coverage;
move template_advanced_literature_review's pytest-cov gate from addopts
to the coverage-report fail_under convention; backfill the
textbook-style local-artifact ignore block into 23 exemplar .gitignores;
add export.toml to template_prose_project and template_methods_paper so
standalone exports ship their infrastructure-dependent script layers.

Docs: disambiguate layered test-timeout defaults in CLAUDE.md; drop the
unverified '300+ files' count from START_HERE.md.

Verified: rendering fallbacks 48 passed; core execution/bounded 64
passed; llm collection clean; ruff + mypy strict clean (1607 files).
Exemplar matrix rerun follows for coverage-provenance refresh.
