+++ START_HERE.md
@@ -224,7 +224,8 @@ tests/              Infrastructure test suite
 projects/templates/ Public canonical exemplars (tracked in git)
 projects/working/   Private working projects (local-only, symlinked from sidecar)
 output/             Generated deliverables (disposable, not committed)
-docs/               Documentation corpus (300+ files, hierarchy in docs/AGENTS.md)
+docs/               Documentation corpus (hierarchy and index in docs/AGENTS.md
+                    and docs/documentation-index.md)
 ```
 
 ---
