+++ TO-DO.md
@@ -39,7 +39,12 @@ These are the prioritized scoped improvements and remaining root-level actions, 
 | `CLEAN-CHECKOUT-MAJ-1` | blocked-external | Medium | Local disposable rehearsal and hosted Linux runner | Run two deterministic hosted-Linux rehearsals and attach the owner/platform receipt; local generated render output is restored only inside the disposable clone after path-boundary validation. | clean-checkout rehearsal receipt | `uv run python scripts/maintenance/release_rehearsal.py --execute --receipt /tmp/template-clean-checkout.json` | non-generated mutation, dirty final tree, changed revision, or unequal deterministic runs must fail |
 | `SECURITY-OWNERSHIP-1` | blocked-external | Medium | Administrator branch-protection and CODEOWNERS receipt | Obtain administrator evidence for required checks, review, force-push protection, and sensitive-path review; local health must remain distinct from authority. | administrator authority receipt | `uv run python scripts/gates/security_scan.py` | repository files or a green local run must not imply remote protection |
 | `SECURITY-PRIVATE-PROMOTION-1` | blocked-external | Medium | Owner authorization, redaction, and export evidence | Obtain an owner-approved private-sidecar promotion record before any promotion; keep the public tree and generated receipts free of private paths and content. | owner promotion receipt | `uv run python scripts/audit/check_tracked_all.py` | private path, sidecar content, credential, or unredacted export must fail public guards |
-| `TEST-ISOLATION-SYSPATH-1` | open | Medium | Root and infra `conftest.py` sys.path loaders | Root-level pytest now uses importlib collection as groundwork (single-tree suites unaffected), but two exemplar trees still cannot share one process: each exemplar's `tests/` is a package named `tests` (first tree's `tests.conftest` wins plugin registration) and every exemplar imports its source as the same top-level `src` package (source-module cross-resolution via `sys.modules`); remaining work is unique exemplar `tests`/`src` package naming; until then keep one exemplar pytest tree per process as documented in `CLAUDE.md`. | isolation restructure record | `uv run pytest projects/templates/template_autopoiesis/tests projects/templates/template_sia/tests --collect-only -q` | two exemplar trees collected in one process must not resolve another tree's modules |
+| `TEST-ISOLATION-SYSPATH-1` | open | Medium | Root and infra `conftest.py` sys.path loaders | Execution-ready recipe (2026-09-06 exploration): adopt the nested unique package per exemplar — move each `src/*.py` module into `src/<exemplar-name>/` while KEEPING the `src/` directory (all layout consumers stay valid: `public_capabilities`, `public_scope`, `project_info`, `working_render`, `export_smoke`, `stage_01_test` discovery) and keeping `src/__init__.py` as an empty namespace shim (the forkability contract is drift-gated by `checks_exemplar.py`); rewrite the ~480 `from src.`/`import src` sites to the unique package; delete all 24 `tests/__init__.py` (importlib groundwork landed); retarget the 11 regression alias loaders (`_PKG_ALIAS` registration) at `src/<name>/__init__.py`; give `checks_exemplar.check_all_export_drift` a package-init resolver; drop `tests/__init__.py` from required files; madlib `pythonpath ["."]`→`["src"]` + helpers→conftest fixtures; methods_paper pinned contract test rewrite; search_project explicit `packages`+`package-dir`; sia/autoresearch pythonpath cleanup; rebase the advanced→literature cross-exemplar symlinks (+1 `../`); note 9 exemplars are already pre-nested and textbook/literature have multi-package `src/` roots (only their flat `from src.` modules move; top-level packages stay). Verify: per-exemplar pytest ×24, the two-tree single-process acceptance below, regression manifest gate, full infra suite, exemplar matrix + provenance refresh. | isolation restructure record | `uv run pytest projects/templates/template_autopoiesis/tests projects/templates/template_sia/tests --collect-only -q` | two exemplar trees collected in one process must not resolve another tree's modules |
+| `CORE-TESTING-REHOME-1` | open | Medium | Established `_`-prefix re-export shim pattern (`security.py`→`_validation.py`) | Re-home the ~16-module pytest/testing cluster (~3,600 lines: `test_runner`, `pytest_orchestration`, `public_matrix_receipt`, `test_performance`, `pytest_profiles`, `script_discovery`, …) from the flat `infrastructure/core/` top level into an `infrastructure/core/testing/` package with one-line re-export shims at the old paths; update AGENTS/SKILL module maps. | re-home record with import-parity check | `uv run python scripts/audit/check_tracked_all.py && uv run pytest tests/infra_tests/core/ -q` | an old import path that stops resolving must fail the parity check |
+| `RENDERING-PANDOC-ARGS-1` | open | Minor | Shared `_output_text.py` dedup landed | Extract the combined-pandoc-args assembly (resource-path triple, formalism filter, pandoc-crossref probe, citeproc+bibliography args) duplicated across `_combined_exports.py` DOCX/EPUB/HTML lanes and `ebook_stage.py` into one builder so the resource-path contract cannot drift per edition. | dedup record | `uv run pytest tests/infra_tests/rendering/ -q` | a lane missing one resource-path leg must fail the parity test |
+| `TEST-MODULE-SPLITS-1` | open | Medium | Section-banner conventions in the eight oversized modules | Split the eight >800-line test modules (test_combined_exports 1,347; test_pipeline 1,333; test_slides_renderer_core 1,213; test_web_renderer 1,067; test_output_validator 980; test_publishing 842; test_counts_doc; test_artifact_manifest_semantics) along their existing section banners into per-production-module files; pure moves, no behavior change. | split record with collection-parity check | `uv run pytest tests/infra_tests/ -q --co` | collected test count before/after the split must match exactly |
+| `RENDERING-LAYERING-1` | open | Medium | Compat-alias re-export pattern | Remove the rendering→publishing layering inversion: `_manuscript_source.py`, `_combined_exports.py`, and `pipeline.py` import `publishing.transmission_bookends`; re-home the `transmission_*` family into a rendering-owned module (or `infrastructure/transmission/`) with publishing-side compat aliases, then decouple the four flat publishing prefix families (`metadata_*`, `release_*`) as subpackages. | layering record | `uv run pytest tests/infra_tests/rendering/ tests/infra_tests/publishing/ -q` | an import-lint rule forbidding rendering→publishing imports must pass |
+| `REHEARSAL-ANALYSIS-EXIT-1` | open | Medium | Hosted rehearsal receipt (run 34062060327) and local probe | Triage the deterministic exit 1 in the fresh-clone release-profile run: active_inference's test stage reports 'All tests passed - ready for analysis' and the stage then exits 1 silently (log ends at the resource-usage line; 2026-09-06 run on main `9c0022745`, pre-dating batch-1). Machinery upgraded 2026-09-06: receipts now carry a bounded redacted output tail on failure (digest-only receipts were undiagnosable), and the plan's matrix command runs the quick profile with two project workers (the serial release profile cannot fit any sane job budget — one exemplar alone exceeded 60 minutes in the local probe). Next: re-dispatch; if the quick-profile rehearsal still exits 1, run the exemplar's post-test analysis stage directly in a disposable clone to surface the swallowed error, fix the stage or its gate, then re-dispatch for CLEAN-CHECKOUT-MAJ-1. The hosted runner's git handles `clone --no-local --revision` (verified: run-1 reached stage_01). | rehearsal failure record + fix | `uv run python scripts/maintenance/release_rehearsal.py --execute --receipt /tmp/rr.json` (exit 0, equal digests) | a silent exit 1 with all tests passing must not be treated as a green rehearsal |
 
 ## Verification order
