+++ infrastructure/rendering/_output_text.py
@@ -9,6 +9,8 @@ errror-message behavior stays byte-identical across formats.
 
 from __future__ import annotations
 
+_ERROR_CONTEXT_LIMIT = 500
+
 
 def _process_output_text(value: bytes | str | None) -> str:
     """Normalize subprocess stdout/stderr values to text for diagnostics."""
@@ -17,3 +19,15 @@ def _process_output_text(value: bytes | str | None) -> str:
     if isinstance(value, str):
         return value
     return ""
+
+
+def _truncate_error_context(stderr_text: str) -> str:
+    """Return bounded stderr/stdout context for ``RenderingError`` messages.
+
+    Shared by the DOCX, EPUB, and MOBI renderers so the bounded-context
+    contract stays byte-identical across formats.
+    """
+    stripped = stderr_text.strip()
+    if not stripped:
+        return "no stderr captured"
+    return stripped[:_ERROR_CONTEXT_LIMIT]
