+++ infrastructure/rendering/docx_renderer.py
@@ -19,12 +19,14 @@ from pathlib import Path
 
 from infrastructure.core.exceptions import RenderingError
 from infrastructure.core.logging.utils import get_logger
-from infrastructure.rendering._output_text import _process_output_text
+from infrastructure.rendering._output_text import (
+    _process_output_text,
+    _truncate_error_context,
+)
 
 logger = get_logger(__name__)
 
 _TIMEOUT_SECONDS = 120
-_ERROR_CONTEXT_LIMIT = 500
 
 
 @dataclass(frozen=True)
@@ -34,14 +36,6 @@ class DocxRenderResult:
     output_path: Path
     size_bytes: int
     duration_seconds: float
-
-
-def _truncate_error_context(stderr_text: str) -> str:
-    """Return bounded stderr/stdout context for RenderingError messages."""
-    stripped = stderr_text.strip()
-    if not stripped:
-        return "no stderr captured"
-    return stripped[:_ERROR_CONTEXT_LIMIT]
 
 
 def render_docx(
@@ -122,7 +116,7 @@ def render_docx(
     except subprocess.TimeoutExpired as exc:
         stderr_text = _process_output_text(exc.stderr) or _process_output_text(exc.stdout) or str(exc)
         raise RenderingError(
-            f"pandoc DOCX render timed out after 120s: {_truncate_error_context(stderr_text)}"
+            f"pandoc DOCX render timed out after {_TIMEOUT_SECONDS}s: {_truncate_error_context(stderr_text)}"
         ) from exc
 
     duration = time.monotonic() - start
