+++ infrastructure/rendering/mobi_renderer.py
@@ -20,12 +20,14 @@ from pathlib import Path
 
 from infrastructure.core.exceptions import RenderingError
 from infrastructure.core.logging.utils import get_logger
-from infrastructure.rendering._output_text import _process_output_text
+from infrastructure.rendering._output_text import (
+    _process_output_text,
+    _truncate_error_context,
+)
 
 logger = get_logger(__name__)
 
 _TIMEOUT_SECONDS = 180  # calibre conversion can be slower than pure pandoc
-_ERROR_CONTEXT_LIMIT = 500
 
 
 @dataclass(frozen=True)
@@ -35,14 +37,6 @@ class MobiRenderResult:
     output_path: Path
     size_bytes: int
     duration_seconds: float
-
-
-def _truncate_error_context(stderr_text: str) -> str:
-    """Return bounded stderr/stdout context for RenderingError messages."""
-    stripped = stderr_text.strip()
-    if not stripped:
-        return "no stderr captured"
-    return stripped[:_ERROR_CONTEXT_LIMIT]
 
 
 def render_mobi(
