+++ projects/templates/template_advanced_literature_review/.gitignore
@@ -8,3 +8,7 @@ __pycache__/
 htmlcov/
 **/*.egg-info/
 .venv/
+# Local artifacts (aligned with template_textbook convention)
+coverage_project.json
+.DS_Store
+.mypy_cache/
