+++ projects/templates/template_autopoiesis/.gitignore
@@ -10,3 +10,6 @@ benchmarks/
 .hypothesis/
 .mypy_cache/
 .ruff_cache/
+# Local artifacts (aligned with template_textbook convention)
+coverage_project.json
+.DS_Store
