+++ projects/templates/template_autoscientists/.gitignore
@@ -7,3 +7,9 @@ __pycache__/
 .venv/
 src/*.egg-info/
 *.egg-info/
+# Local artifacts (aligned with template_textbook convention)
+htmlcov/
+coverage_project.json
+.DS_Store
+.mypy_cache/
+.ruff_cache/
