+++ projects/templates/template_literature_meta_analysis/.gitignore
@@ -15,3 +15,9 @@ __pycache__/
 # exemplar evidence. Commit derived assessments and provenance, never raw text.
 output/fulltext/*
 !output/fulltext/fulltext_inventory.json
+# Local artifacts (aligned with template_textbook convention)
+htmlcov/
+coverage_project.json
+.DS_Store
+.mypy_cache/
+.ruff_cache/
