+++ projects/templates/template_methods_paper/export.toml
@@ -1,0 +1,2 @@
+[template_export]
+cross_root_dependencies = ["infrastructure"]
