+++ projects/templates/template_prose_project/export.toml
@@ -1,0 +1,2 @@
+[template_export]
+cross_root_dependencies = ["infrastructure"]
