+++ projects/templates/template_redacted_report/.gitignore
@@ -14,3 +14,7 @@ __pycache__/
 htmlcov/
 coverage_project.json
 .coverage*
+# Local artifacts (aligned with template_textbook convention)
+.DS_Store
+.mypy_cache/
+.ruff_cache/
