+++ projects/templates/template_search_project/.gitignore
@@ -12,3 +12,9 @@ __pycache__/
 .env
 .env.local
 output/search/cache/
+# Local artifacts (aligned with template_textbook convention)
+htmlcov/
+coverage_project.json
+.DS_Store
+.mypy_cache/
+.ruff_cache/
