+++ tests/infra_tests/llm/conftest.py
@@ -1,3 +1,4 @@
+import os
 import threading
 
 import pytest
@@ -139,3 +140,169 @@ def patch_llm_client_for_tests(request, monkeypatch):
     monkeypatch.setenv("OLLAMA_MODEL", "gemma3:4b")
     # Disable auto system prompt injection for predictable test behavior
     monkeypatch.setenv("LLM_AUTO_INJECT_SYSTEM_PROMPT", "false")
+
+
+# ============================================================================
+# Ollama (requires_ollama tests)
+# ============================================================================
+
+
+def _env_truthy(name: str) -> bool:
+    return os.environ.get(name, "").strip().lower() in ("1", "true", "yes")
+
+
+def _ollama_test_pull_timeout() -> float | None:
+    # Default 180s (3 min) — long enough for a small model on a warm cache,
+    # short enough that a missing/slow registry fails fast with a clear
+    # message pointing to manual `ollama pull <model>`. Override with
+    # OLLAMA_TEST_PULL_TIMEOUT (seconds, or "none"/"inf" to disable).
+    raw = os.environ.get("OLLAMA_TEST_PULL_TIMEOUT", "180").strip()
+    if not raw or raw.lower() in ("none", "inf"):
+        return None
+    return float(raw)
+
+
+@pytest.fixture(scope="session")
+def ensure_ollama_for_tests():
+    """Ensure Ollama is running and a small/fast model is available for tests.
+
+    - Starts Ollama when possible (same as :func:`ensure_ollama_ready`).
+    - When ``OLLAMA_SKIP_TEST_MODEL_PULL`` is unset/false, runs ``ollama pull`` for
+      ``OLLAMA_TEST_PULL_MODEL`` (default ``smollm2``) if no matching small model is installed.
+    - Set ``OLLAMA_SKIP_TEST_MODEL_PULL=1`` for air-gapped runs (install models manually).
+
+    See ``tests/README.md`` for environment variables.
+    """
+    from infrastructure.core.logging.utils import get_logger
+    from infrastructure.llm.utils.ollama import (
+        ensure_ollama_ready,
+        get_model_names,
+        is_ollama_running,
+        pull_ollama_model,
+        small_fast_preference_matches,
+    )
+
+    logger = get_logger(__name__)
+    skip_pull = _env_truthy("OLLAMA_SKIP_TEST_MODEL_PULL")
+    pull_model = (os.environ.get("OLLAMA_TEST_PULL_MODEL") or "smollm2").strip()
+    pull_timeout = _ollama_test_pull_timeout()
+
+    if is_ollama_running():
+        logger.info("✓ Ollama server is already running")
+    else:
+        logger.warning("⚠️  Ollama server is not running - attempting to start...")
+        if not ensure_ollama_ready(auto_start=True):
+            pytest.fail(
+                "\n" + "=" * 80 + "\n"
+                "❌ CRITICAL: Ollama server cannot be started for tests!\n"
+                "=" * 80 + "\n"
+                "Tests require a working Ollama installation.\n\n"
+                "Troubleshooting:\n"
+                "  1. Install Ollama: https://ollama.ai\n"
+                "  2. Start Ollama manually: ollama serve\n"
+                "  3. Verify installation: ollama --version\n"
+                "  4. Check if port 11434 is available: lsof -i :11434\n"
+                "  5. Pull a small test model: ollama pull smollm2\n\n"
+                "To deselect Ollama tests: pytest -m 'not requires_ollama'\n"
+                "=" * 80
+            )
+
+    def _fail_no_models() -> None:
+        pytest.fail(
+            "\n" + "=" * 80 + "\n"
+            "❌ CRITICAL: Ollama server is running but has no models!\n"
+            "=" * 80 + "\n"
+            "Tests require at least one Ollama model.\n\n"
+            "Quick install for tests:\n"
+            "  ollama pull smollm2\n\n"
+            "Or set OLLAMA_SKIP_TEST_MODEL_PULL=0 (default) to allow auto-pull.\n"
+            "Air-gapped: pre-install smollm2 and set OLLAMA_SKIP_TEST_MODEL_PULL=1.\n\n"
+            "Verify: ollama list\n"
+            "=" * 80
+        )
+
+    models = get_model_names()
+
+    def _maybe_pull(reason: str) -> None:
+        nonlocal models
+        if skip_pull:
+            logger.warning(
+                "Ollama has no small/fast test model (%s); OLLAMA_SKIP_TEST_MODEL_PULL is set — "
+                "tests may use slower fallback models or fail if no model can be loaded. "
+                "Install e.g. smollm2 manually.",
+                reason,
+            )
+            return
+        logger.info("Pulling test model %r (%s)...", pull_model, reason)
+        ok, err = pull_ollama_model(pull_model, timeout=pull_timeout)
+        if not ok:
+            pytest.fail(
+                "\n" + "=" * 80 + "\n"
+                "❌ CRITICAL: ollama pull failed for test model.\n"
+                "=" * 80 + "\n"
+                f"Model: {pull_model}\n"
+                f"Timeout: {pull_timeout}s (override with OLLAMA_TEST_PULL_TIMEOUT)\n"
+                f"Error: {err}\n\n"
+                "Fastest fix — pull the model manually in another terminal:\n"
+                f"  ollama pull {pull_model}\n\n"
+                "Then re-run tests, or set OLLAMA_SKIP_TEST_MODEL_PULL=1 to\n"
+                "use whatever is already installed (see `ollama list`).\n"
+                "=" * 80
+            )
+        models = get_model_names()
+        if not models:
+            _fail_no_models()
+        if not small_fast_preference_matches(models):
+            pytest.fail(
+                "\n" + "=" * 80 + "\n"
+                "❌ CRITICAL: After pull, no small/fast preference model is listed.\n"
+                "=" * 80 + "\n"
+                f"Expected one of the preferences matching: {pull_model!r}\n"
+                f"Got: {models!r}\n"
+                "=" * 80
+            )
+
+    if not models:
+        _maybe_pull("no models installed")
+    elif not small_fast_preference_matches(models):
+        _maybe_pull("no smollm2/gemma2:2b/… style model in library")
+
+    if not models:
+        _fail_no_models()
+
+    # Capability probe: a listed model proves nothing about the runtime — a
+    # broken install (e.g. Homebrew Ollama whose bundled llama-server binary is
+    # missing) lists models fine and then returns HTTP 500 on every generate.
+    # Probe actual inference once per session so 20+ LLM tests fail with one
+    # actionable diagnosis instead of identical confusing per-test errors.
+    from infrastructure.llm.utils.models import preload_model, select_small_fast_model
+
+    probe_model = select_small_fast_model() or models[0]
+    probe_ok, probe_err = preload_model(probe_model, timeout=120.0, retries=1)
+    if not probe_ok:
+        pytest.fail(
+            "\n" + "=" * 80 + "\n"
+            "❌ CRITICAL: Ollama lists models but cannot run inference!\n"
+            "=" * 80 + "\n"
+            f"Probe model: {probe_model}\n"
+            f"Error: {probe_err}\n\n"
+            "The server is up and models are installed, but loading a model failed —\n"
+            "this usually means a broken Ollama runtime (e.g. a Homebrew upgrade that\n"
+            "left the llama-server binary missing).\n\n"
+            "Troubleshooting:\n"
+            "  1. ollama run " + probe_model + " 'hi'   # reproduce directly\n"
+            "  2. brew reinstall ollama && brew services restart ollama\n"
+            "  3. Or reinstall from https://ollama.ai\n\n"
+            "To deselect Ollama tests: pytest -m 'not requires_ollama'\n"
+            "=" * 80
+        )
+
+    logger.info("✓ Ollama ready for tests with %d model(s): %s", len(models), ", ".join(models[:5]))
+    return True
+
+
+@pytest.fixture(autouse=True)
+def _ollama_requires_session(request: pytest.FixtureRequest) -> None:
+    """Run session-scoped Ollama ensure once for any ``requires_ollama`` test."""
+    if request.node.get_closest_marker("requires_ollama"):
+        request.getfixturevalue("ensure_ollama_for_tests")
